+++ src/index.ts
@@ -1,11 +1,10 @@
 export { DynamicClusterPool } from './pools/cluster/dynamic'
 export { FixedClusterPool } from './pools/cluster/fixed'
 export type { ClusterPoolOptions } from './pools/cluster/fixed'
-export type { IPool, PoolOptions } from './pools/pool'
+export type { IPool, PoolEmitter, PoolOptions } from './pools/pool'
 export type {
   ErrorHandler,
   ExitHandler,
-  IPoolWorker,
   MessageHandler,
   OnlineHandler
 } from './pools/pool-worker'
@@ -14,7 +13,6 @@ export type { WorkerChoiceStrategy } from './pools/selection-strategies/selectio
 export { DynamicThreadPool } from './pools/thread/dynamic'
 export { FixedThreadPool } from './pools/thread/fixed'
 export type { ThreadWorkerWithMessageChannel } from './pools/thread/fixed'
-export { AbstractWorker } from './worker/abstract-worker'
 export { ClusterWorker } from './worker/cluster-worker'
 export { ThreadWorker } from './worker/thread-worker'
 export { KillBehaviors } from './worker/worker-options'